t options: sort by last modification date

+++ srcs/ft_btreeadd_ls.c
@@ -6,14 +6,14 @@
 /*   By: mnunnari <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/05/17 19:27:42 by mnunnari          #+#    #+#             */
-/*   Updated: 2017/05/17 19:33:17 by mnunnari         ###   ########.fr       */
+/*   Updated: 2017/05/18 16:42:23 by mnunnari         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 void	ft_btreeadd_ls(t_btree **root, t_btree *new, char *opts,
-					int (*cmpf)(t_btree*, t_btree*, char*))
+					long (*cmpf)(t_btree*, t_btree*, char*))
 {
 	if (*root == NULL)
 		*root = new;
